chore: migrate login to authentication api
