+++ dream-big-api/app/controllers/application_controller.rb
@@ -1,13 +1,5 @@
 class ApplicationController < ActionController::API
-  include JsonWebToken
-
-  before_action :authenticate_request
-
-  private
-      def authenticate_request
-        header = request.headers["Authorization"]
-        header = header.split(" ").last if header
-        decoded = jwt_decode(header)
-        @current_user = User.find(decoded[:user_id])
-      end
+  def headers
+    request.headers
+  end
 end
